+++ projects/OG-Analytics/src/main/java/com/opengamma/analytics/financial/model/volatility/surface/BasisSplineVolatilitySurfaceProvider.java
@@ -3,11 +3,10 @@
  * 
  * Please see distribution for license.
  */
-package com.opengamma.analytics.financial.interestrate.capletstripping;
+package com.opengamma.analytics.financial.model.volatility.surface;
 
 import java.util.List;
 
-import com.opengamma.analytics.financial.model.volatility.surface.VolatilitySurface;
 import com.opengamma.analytics.math.function.Function1D;
 import com.opengamma.analytics.math.function.Function2D;
 import com.opengamma.analytics.math.interpolation.BasisFunctionAggregation;
