+++ projects/OG-Analytics/src/main/java/com/opengamma/analytics/financial/model/volatility/surface/VolatilitySurfaceProvider.java
@@ -3,9 +3,8 @@
  * 
  * Please see distribution for license.
  */
-package com.opengamma.analytics.financial.interestrate.capletstripping;
+package com.opengamma.analytics.financial.model.volatility.surface;
 
-import com.opengamma.analytics.financial.model.volatility.surface.VolatilitySurface;
 import com.opengamma.analytics.math.matrix.DoubleMatrix1D;
 import com.opengamma.analytics.math.surface.Surface;
