+++ projects/OG-Analytics/src/main/java/com/opengamma/analytics/math/function/ParameterizedCurveVectorFunction.java
@@ -3,10 +3,8 @@
  * 
  * Please see distribution for license.
  */
-package com.opengamma.analytics.financial.interestrate.capletstripping;
+package com.opengamma.analytics.math.function;
 
-import com.opengamma.analytics.math.function.Function1D;
-import com.opengamma.analytics.math.function.ParameterizedCurve;
 import com.opengamma.analytics.math.matrix.DoubleMatrix1D;
 import com.opengamma.analytics.math.matrix.DoubleMatrix2D;
 import com.opengamma.util.ArgumentChecker;
