+++ projects/OG-Analytics/src/main/java/com/opengamma/analytics/math/function/ParameterizedCurveVectorFunctionProvider.java
@@ -3,9 +3,8 @@
  * 
  * Please see distribution for license.
  */
-package com.opengamma.analytics.financial.interestrate.capletstripping;
+package com.opengamma.analytics.math.function;
 
-import com.opengamma.analytics.math.function.ParameterizedCurve;
 import com.opengamma.util.ArgumentChecker;
 
 /**
