+++ projects/OG-Analytics/src/main/java/com/opengamma/analytics/math/minimization/PositiveOrZero.java
@@ -1,4 +1,4 @@
-package com.opengamma.analytics.financial.interestrate.capletstripping;
+package com.opengamma.analytics.math.minimization;
 
 import com.opengamma.analytics.math.function.Function1D;
 import com.opengamma.analytics.math.matrix.DoubleMatrix1D;
