+++ projects/OG-Analytics/src/test/java/com/opengamma/analytics/financial/interestrate/capletstripping/SABRDirectTest.java
@@ -13,15 +13,15 @@ import org.testng.annotations.Test;
 
 import com.opengamma.analytics.financial.interestrate.capletstripping.CapletStrippingImp;
 import com.opengamma.analytics.financial.interestrate.capletstripping.CapletStrippingResult;
-import com.opengamma.analytics.financial.interestrate.capletstripping.DiscreteVolatilityFunction;
-import com.opengamma.analytics.financial.interestrate.capletstripping.DiscreteVolatilityFunctionProvider;
 import com.opengamma.analytics.financial.interestrate.capletstripping.MultiCapFloorPricer;
-import com.opengamma.analytics.financial.interestrate.capletstripping.ParameterizedSABRModelDiscreteVolatilityFunctionProvider;
-import com.opengamma.analytics.financial.interestrate.capletstripping.VectorFunction;
 import com.opengamma.analytics.financial.model.interestrate.curve.ForwardCurve;
+import com.opengamma.analytics.financial.model.volatility.discrete.DiscreteVolatilityFunction;
+import com.opengamma.analytics.financial.model.volatility.discrete.DiscreteVolatilityFunctionProvider;
+import com.opengamma.analytics.financial.model.volatility.discrete.ParameterizedSABRModelDiscreteVolatilityFunctionProvider;
 import com.opengamma.analytics.math.curve.InterpolatedDoublesCurve;
 import com.opengamma.analytics.math.differentiation.VectorFieldFirstOrderDifferentiator;
 import com.opengamma.analytics.math.function.Function1D;
+import com.opengamma.analytics.math.function.VectorFunction;
 import com.opengamma.analytics.math.interpolation.CombinedInterpolatorExtrapolatorFactory;
 import com.opengamma.analytics.math.interpolation.Interpolator1DFactory;
 import com.opengamma.analytics.math.interpolation.PenaltyMatrixGenerator;
