+++ projects/OG-Analytics/src/test/java/com/opengamma/analytics/financial/interestrate/capletstripping/SABRInterpolatedTermStructureTest.java
@@ -13,6 +13,9 @@ import org.testng.annotations.Test;
 import com.opengamma.analytics.financial.model.volatility.smile.function.SABRFormulaData;
 import com.opengamma.analytics.financial.model.volatility.smile.function.SABRHaganVolatilityFunction;
 import com.opengamma.analytics.financial.provider.description.interestrate.MulticurveProviderDiscount;
+import com.opengamma.analytics.math.function.ConcatenatedVectorFunction;
+import com.opengamma.analytics.math.function.InterpolatedVectorFunctionProvider;
+import com.opengamma.analytics.math.function.VectorFunction;
 import com.opengamma.analytics.math.interpolation.CombinedInterpolatorExtrapolatorFactory;
 import com.opengamma.analytics.math.interpolation.Interpolator1D;
 import com.opengamma.analytics.math.interpolation.Interpolator1DFactory;
