+++ projects/OG-Analytics/src/test/java/com/opengamma/analytics/financial/interestrate/capletstripping/demo/CapletStrippingPSplineTermStrucutreDemo.java
@@ -9,7 +9,6 @@ import java.util.Arrays;
 
 import org.testng.annotations.Test;
 
-import com.opengamma.analytics.financial.interestrate.capletstripping.BasisSplineVolatilityTermStructureProvider;
 import com.opengamma.analytics.financial.interestrate.capletstripping.CapletStripper;
 import com.opengamma.analytics.financial.interestrate.capletstripping.CapletStripperPSplineTermStructure;
 import com.opengamma.analytics.financial.interestrate.capletstripping.CapletStrippingResult;
@@ -17,6 +16,7 @@ import com.opengamma.analytics.financial.interestrate.capletstripping.CapletStri
 import com.opengamma.analytics.financial.interestrate.capletstripping.CombinedCapletStrippingResults;
 import com.opengamma.analytics.financial.interestrate.capletstripping.MarketDataType;
 import com.opengamma.analytics.financial.interestrate.capletstripping.MultiCapFloorPricer;
+import com.opengamma.analytics.financial.model.volatility.surface.BasisSplineVolatilityTermStructureProvider;
 import com.opengamma.analytics.math.matrix.DoubleMatrix1D;
 
 /**
