+++ projects/OG-Analytics/src/test/java/com/opengamma/analytics/financial/interestrate/capletstripping/demo/SABRInterpolatedTermStructureDemo.java
@@ -17,15 +17,15 @@ import com.opengamma.analytics.financial.interestrate.capletstripping.CapletStri
 import com.opengamma.analytics.financial.interestrate.capletstripping.CapletStripperSmileModel;
 import com.opengamma.analytics.financial.interestrate.capletstripping.CapletStrippingResult;
 import com.opengamma.analytics.financial.interestrate.capletstripping.CapletStrippingSetup;
-import com.opengamma.analytics.financial.interestrate.capletstripping.ConcatenatedVectorFunction;
-import com.opengamma.analytics.financial.interestrate.capletstripping.InterpolatedVectorFunctionProvider;
 import com.opengamma.analytics.financial.interestrate.capletstripping.MarketDataType;
 import com.opengamma.analytics.financial.interestrate.capletstripping.MultiCapFloorPricer;
-import com.opengamma.analytics.financial.interestrate.capletstripping.VectorFunction;
 import com.opengamma.analytics.financial.model.volatility.smile.function.SABRFormulaData;
 import com.opengamma.analytics.financial.model.volatility.smile.function.SABRHaganVolatilityFunction;
 import com.opengamma.analytics.financial.provider.description.interestrate.MulticurveProviderDiscount;
 import com.opengamma.analytics.math.curve.InterpolatedDoublesCurve;
+import com.opengamma.analytics.math.function.ConcatenatedVectorFunction;
+import com.opengamma.analytics.math.function.InterpolatedVectorFunctionProvider;
+import com.opengamma.analytics.math.function.VectorFunction;
 import com.opengamma.analytics.math.interpolation.CombinedInterpolatorExtrapolatorFactory;
 import com.opengamma.analytics.math.interpolation.Interpolator1D;
 import com.opengamma.analytics.math.interpolation.Interpolator1DFactory;
